backend/requirements.txt
Flask
Flask-CORS
numpy
opencv-python
faiss-cpu
insightface==0.7.2
onnxruntime
onnx
scikit-image
scipy
easydict
firebase-admin 
